NUCLEO_L496ZG JSON Clock Configuration
